% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cens_auth}
\alias{cens_auth}
\title{Authorize use of the Census API}
\usage{
cens_auth()
}
\value{
a Census API key
}
\description{
Tries environment variables \code{CENSUS_API_KEY} and \code{CENSUS_KEY}, in that order.
If none is found and R is used in interactive mode, will prompt the user for
a key.
}
