% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easycensus-package.R
\docType{package}
\name{easycensus-package}
\alias{easycensus}
\alias{easycensus-package}
\title{easycensus: Quickly Extract and Marginalize U.S. Census Tables}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Extracting desired data using the proper Census variable names can be time-consuming. This package takes the pain out of that process.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://corymccartan.github.io/easycensus/}
  \item Report bugs at \url{https://github.com/CoryMcCartan/easycensus/issues}
}

}
\author{
\strong{Maintainer}: Cory McCartan \email{cmccartan@g.harvard.edu}

}
\keyword{internal}
