% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{get_est}
\alias{get_est}
\alias{get_se}
\alias{get_moe}
\title{Extract estimates, standard errors, and margins of error}
\usage{
get_est(x)

get_se(x)

get_moe(x, conf = 0.9)
}
\arguments{
\item{x}{An \link{estimate} vector.}

\item{conf}{The confidence level to use in constructing the margin of error.}
}
\value{
An \link{estimate} vector.
}
\description{
Getter functions for \code{\link[=estimate]{estimate()}} vectors.
}
\examples{
x = estimate(1, 0.1)
get_est(x)
get_moe(x)

}
