% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_table.R
\name{find_dec_table}
\alias{find_dec_table}
\title{Find a decennial census table with variables of interest}
\usage{
find_dec_table(..., show = 2)
}
\arguments{
\item{...}{Variables to look for. These can be length-1 character vectors,
or, for convenience, can be left unquoted (see examples).}

\item{show}{How many matching tables to show. Increase this to show more
possible matches, at the cost of more output.}
}
\value{
The codes for the top \code{show} tables, invisibly.
}
\description{
This function uses fuzzy matching to help identify SF1 tables from the census
which contain variables of interest. Matched table codes are printed out,
along with the Census-provided table description, the parsed variable names,
and example table cells.
}
\examples{
find_dec_table("sex", "age")
find_dec_table(tenure, race, show=1)

}
