% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{format.estimate}
\alias{format.estimate}
\title{Format an estimate}
\usage{
\method{format}{estimate}(x, conf = 0.9, digits = 2, trim = FALSE, ..., formatter = fmt_plain)
}
\arguments{
\item{x}{An \link{estimate} vector}

\item{digits}{The number of dig}
}
\description{
Format an estimate for pretty printing
}
