% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_table.R
\name{get_dec_table}
\alias{get_dec_table}
\title{Download data from a decennial census table}
\usage{
get_dec_table(geography, table, ..., drop_total = FALSE)
}
\arguments{
\item{geography}{The geography level to download data for. Usually one of
\code{state}, \code{county}, \code{tract}, \verb{block group}, \code{block}, \code{zcta}, etc.}

\item{table}{The table code to download. See \code{\link[=find_dec_table]{find_dec_table()}} for help
identifying a table of interest. Note: some tables are split into
A/B/C/etc. versions by race; this function unifies all of these tables
under one code. So, for example, use \code{P012}, not \code{P012A}.}

\item{...}{Arguments passed to \code{\link[tidycensus:get_decennial]{tidycensus::get_decennial()}}, e.g. \code{year},
\code{state}, \code{county}, \code{geometry}.}

\item{drop_total}{Whether to filter out variables which are totals across
another variable. Recommended only after inspection of the underlying
table.}
}
\value{
A tibble of decennial Census data in tidy format, with columns
\code{GEOID}, \code{NAME}, \code{variable} (containing the Census variable code), \code{value},
and additional factor columns specific to the table.
}
\description{
Download data from a decennial census table
}
\examples{
get_dec_table("state", "P003")
get_dec_table("state", "H002")
get_dec_table("state", "H002", drop_total=TRUE)

}
