% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_census_years.R
\name{merge_census_years}
\alias{merge_census_years}
\title{Merge Census Data Across Years}
\usage{
merge_census_years(..., id_col = "tract_id")
}
\arguments{
\item{...}{Two or more data frames to merge.}

\item{id_col}{The column name that identifies each area (default is "tract_id").}
}
\value{
A merged data frame containing all years.
}
\description{
Combines two or more census datasets by a common ID column (like "tract_id").
This helps you compare variables across different years.
}
\examples{
\dontrun{
df1 <- data.frame(tract_id = 1:2, income_2010 = c(50000, 55000))
df2 <- data.frame(tract_id = 1:2, income_2020 = c(60000, 65000))
merge_census_years(df1, df2, id_col = "tract_id")
}
}
