% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{moe_product}
\alias{moe_prop}
\alias{moe_ratio}
\alias{moe_sum}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{tidycensus}{\code{\link[tidycensus]{moe_product}}, \code{\link[tidycensus]{moe_prop}}, \code{\link[tidycensus]{moe_ratio}}, \code{\link[tidycensus]{moe_sum}}}
}}

