% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_ltdb_vars.R
\name{rename_ltdb_vars}
\alias{rename_ltdb_vars}
\title{Rename LTDB variable names}
\usage{
rename_ltdb_vars(df)
}
\arguments{
\item{df}{A data frame with LTDB-style column names.}
}
\value{
A data frame with renamed columns.
}
\description{
Renames hard-to-read variable names from LTDB or ACS data to clearer ones.
}
\examples{
\dontrun{
df <- data.frame(povrate = 10, medinc = 55000)
rename_ltdb_vars(df)
}
}
