% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_table.R
\docType{data}
\name{tables}
\alias{tables}
\alias{tables_sf1}
\alias{tables_acs}
\title{Parsed Census SF1 and ACS Tables}
\format{
A list of \code{cens_table} objects, which are just lists with four elements:
\itemize{
\item \code{concept}, a human-readable name
\item \code{tables}, the constituent table codes
\item \code{surveys}, the supported surveys
\item \code{dims}, the parsed names of the dimensions of the tables
\item \code{vars}, a \code{tibble} with all of the parsed variable values
}

An object of class \code{list} of length 83.

An object of class \code{list} of length 848.
}
\usage{
tables_sf1

tables_acs
}
\description{
Contains parsed table information for the 2010 Decennial Summary File 1 and
2019 ACS 5-year and 1-year tables.
This parsed information is used internally in \code{\link[=cens_find_dec]{cens_find_dec()}},
\code{\link[=cens_find_acs]{cens_find_acs()}}, \code{\link[=cens_get_dec]{cens_get_dec()}}, and \code{\link[=cens_get_acs]{cens_get_acs()}}.
For other sets of tables, try using \code{\link[=cens_parse_tables]{cens_parse_tables()}}.
}
\keyword{datasets}
