% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easycensus-package.R
\docType{package}
\name{easycensus-package}
\alias{easycensus}
\alias{easycensus-package}
\title{easycensus: Quickly Extract and Marginalize U.S. Census Tables}
\description{
Extracting desired data using the proper Census variable names can be time-consuming. This package takes the pain out of that process.
}
\author{
\strong{Maintainer}: Cory McCartan \email{cmccartan@g.harvard.edu}

}
\keyword{internal}
