% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easycensus-package.R
\docType{package}
\name{easycensus-package}
\alias{easycensus}
\alias{easycensus-package}
\title{easycensus: Quickly Find, Extract, and Marginalize U.S. Census Tables}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Extracting desired data using the proper Census variable names can be time-consuming. This package takes the pain out of that process by providing functions to quickly locate variables and download labeled tables from the Census APIs (\url{https://www.census.gov/data/developers/data-sets.html}).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://corymccartan.com/easycensus/}
  \item \url{https://github.com/CoryMcCartan/easycensus/}
  \item Report bugs at \url{https://github.com/CoryMcCartan/easycensus/issues}
}

}
\author{
\strong{Maintainer}: Cory McCartan \email{cmccartan@g.harvard.edu}

}
\keyword{internal}
