% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{format.estimate}
\alias{format.estimate}
\title{Format an estimate}
\usage{
\method{format}{estimate}(x, conf = 0.9, digits = 2, trim = FALSE, ..., formatter = fmt_plain)
}
\arguments{
\item{x}{An \link{estimate} vector}

\item{conf}{The confidence level to use in converting the margin of error to
a standard error. Defaults to 90\%, which is what the Census Bureau uses for
ACS estimates.}

\item{digits}{The number of dig}

\item{trim}{logical; if \code{FALSE}, logical, numeric and complex values are
right-justified to a common width: if \code{TRUE} the leading blanks for
justification are suppressed.}

\item{...}{Ignored.}

\item{formatter}{the formatting function to use internally}
}
\description{
Format an estimate for pretty printing
}
